% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_in_header.R
\name{rmd_in_header}
\alias{rmd_in_header}
\title{Generate an html file to be included as `in_header` for Rmarkdown documents}
\usage{
rmd_in_header(header_file = "cssloaders_in_header.html",
  type = getOption("spinner.type", default = 1),
  color = getOption("spinner.color", default = "#0275D8"),
  size = getOption("spinner.size", default = 1),
  color.background = getOption("spinner.color.background"))
}
\arguments{
\item{header_file}{The path to the html header file to be created}

\item{type}{The type of spinner to use, valid values are integers between 1-8. They correspond to the enumeration in \url{https://projects.lukehaas.me/css-loaders/}}

\item{color}{The color of the spinner to be applied in HTML in hex format}

\item{size}{The size of the spinner, relative to it's default size.}

\item{color.background}{For certain spinners (type 2-3), you will need to specify the background colour of the spinner}
}
\description{
Generate an html file to be included as `in_header` for Rmarkdown documents
}
