% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinycssloaders-package.r
\docType{package}
\name{shinycssloaders}
\alias{shinycssloaders}
\title{shinycssloaders}
\description{
Add automatic loading animations to your Shiny outputs (e.g. plots, tables) by adding a single function call
around the Shiny output function. Animations automatically show/hide when the output
is recalculating/recalculated. You can use your own custom image, or choose from 8 different types of
built-in animations, as well customize the color and size of the built-in animations.
}
\details{
For further reference on how to use the package, please refer to \url{https://github.com/daattali/shinycssloaders}
}
