% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/withSpinner.R
\name{withCustomSpinner}
\alias{withCustomSpinner}
\title{Add a custom spinner image when an output is recalculating}
\usage{
withCustomSpinner(
  ui_element,
  image,
  height = NULL,
  width = NULL,
  proxy.height = NULL,
  id = NULL
)
}
\arguments{
\item{ui_element}{A UI element that should be wrapped with a spinner when the corresponding output is being calculated.}

\item{image}{The path or URL of the image to use.}

\item{height}{The height for the spinner, in pixels. If not provided, then the original
size of the image is used.}

\item{width}{The width for the spinner, in pixels. If not provided, then the original
size of the image is used.}

\item{proxy.height}{If the output doesn't specify the output height, you can set a proxy height. It defaults to "400px" for outputs with undefined height.}

\item{id}{The HTML ID to use for the spinner. If you don't provide one, it will be generated automatically.}
}
\description{
Similar to \code{\link{withSpinner}}, but a custom image can be provided.
}
