\name{shinycssloaders-package}
\alias{shinycssloaders-package}
\alias{shinycssloaders}
\docType{package}
\title{
\packageTitle{shinycssloaders}
}
\description{
\packageDescription{shinycssloaders}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{shinycssloaders}
\packageIndices{shinycssloaders}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{shinycssloaders}

Maintainer: \packageMaintainer{shinycssloaders}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
