% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinycssloaders-package.r
\docType{package}
\name{shinycssloaders}
\alias{shinycssloaders}
\alias{shinycssloaders-package}
\title{shinycssloaders.}
\description{
Add automatic loading animations to your Shiny outputs (e.g. plots, tables) by adding a single function call 
around the Shiny output function. Animations automatically show / hide when the output 
is recalculating / recalculated (or error is throws). You can choose from 8 different types of animations and you can
also customize the color / size of the animation.
}
\details{
For further reference on how to use the package, please refer to \url{https://github.com/andrewsali/shinycssloaders}
}
