% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pageSpinner.R
\name{showHidePage}
\alias{showHidePage}
\alias{showPageSpinner}
\alias{hidePageSpinner}
\title{Show/hide a full-page spinner}
\usage{
showPageSpinner(expr)

hidePageSpinner(id)
}
\arguments{
\item{expr}{(optional) An R expression to run while showing the spinner. The
spinner will automatically get hidden when this expression completes.}
}
\value{
If \code{expr} is provided, the result of \code{expr} is returned. Otherwise, \code{NULL}.
}
\description{
Use these functions to show or hide the full-page spinner. The spinner must first be
created in the Shiny app's UI using \code{\link[=pageSpinner]{pageSpinner()}}.
}
\examples{
if (interactive()) {
  library(shiny)

  #--- Example 1: Using showPageSpinner/hidePageSpinner ---

  ui <- fluidPage(
    pageSpinner(),
    actionButton("go", "Go"),
    plotOutput("plot")
  )
  server <- function(input, output) {
    observeEvent(input$go, {
      showPageSpinner()
      Sys.sleep(1)
      hidePageSpinner()
    })
    output$plot <- renderPlot({
     plot(runif(10))
    })
  }
  shinyApp(ui, server)

  #--- Example 2: Using showPageSpinner with expr ---

  some_slow_function <- function() {
    Sys.sleep(1)
  }

  ui <- fluidPage(
    pageSpinner(),
    actionButton("go", "Go"),
    plotOutput("plot")
  )
  server <- function(input, output) {
    observeEvent(input$go, {
      showPageSpinner({ some_slow_function() })
    })
    output$plot <- renderPlot({
     plot(runif(10))
    })
  }
  shinyApp(ui, server)
}
}
\seealso{
\code{\link[=pageSpinner]{pageSpinner()}}
}
