% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/withSpinner.R
\name{withSpinner}
\alias{withSpinner}
\title{Add a spinner (loader) that shows when an output is recalculating}
\usage{
withSpinner(
  ui_element,
  type = getOption("spinner.type", default = 1),
  color = getOption("spinner.color", default = "#0275D8"),
  size = getOption("spinner.size", default = 1),
  color.background = getOption("spinner.color.background"),
  custom.css = FALSE,
  proxy.height = NULL,
  id = NULL
)
}
\arguments{
\item{ui_element}{A UI element that should be wrapped with a spinner when the corresponding output is being calculated.}

\item{type}{The type of spinner to use, valid values are integers between 0-8 (0 means no spinner). Check out \url{https://daattali.com/shiny/shinycssloaders-demo} to see the different types of spinners.}

\item{color}{The color of the spinner in hex format}

\item{size}{The size of the spinner, relative to its default size.}

\item{color.background}{For certain spinners (type 2-3), you will need to specify the background color of the spinner}

\item{custom.css}{Set to \code{TRUE} if you have your own custom CSS that you defined and you don't want the automatic CSS applied to the spinner.}

\item{proxy.height}{If the output doesn't specify the output height, you can set a proxy height. It defaults to "400px" for outputs with undefined height.}

\item{id}{The HTML ID to use for the spinner. If you don't provide one, it will be generated automatically.}
}
\description{
Add a spinner (loader) that shows when an output is recalculating
}
\examples{
\dontrun{withSpinner(plotOutput("my_plot"))}
}
